chore(clippy): fix warnings across V31 workspace.
- zion-pool: remove unused imports, unused `last_job_fetch`, deprecated
  `set_linger`, unused pattern variables, redundant closure, field reassign,
  and suppress too_many_arguments on `run_bridge_task`.
- zion-miner: remove unused `AtomicPtr` import, rename Windows
  `PVECTORED_EXCEPTION_HANDLER` to CamelCase.
- zion-dao: remove unused `ApiOk` type.
- zion-multichain: suppress too_many_arguments on `dfs_paths`.
- cosmic-harmony / cosmic-harmony-v3: collapse nested else-if.

All workspace tests and clippy pass.

+++ V31/L1/pool/src/vardiff.rs
@@ -151,7 +151,7 @@ pub fn difficulty_to_target(difficulty: u64) -> [u8; 32] {
 
     // Number of leading zero bits = floor(log2(difficulty)).
     // For difficulty D, target ≈ 2^256 / D = 2^(256 - log2(D)).
-    let leading_zeros = (difficulty as u64).leading_zeros() as usize;
+    let leading_zeros = difficulty.leading_zeros() as usize;
     let log2_d = 63 - leading_zeros; // floor(log2(difficulty)) = position of MSB
     let shift_bits = log2_d.min(256);
